Start introducing typescript
