+++ static/.eslintignore
@@ -2,3 +2,6 @@ libs
 *.out.js
 *.map
 *.json
+*.tsx
+*.css
+*.png
