+++ static/rustw.ts
