+++ webpack.config.js
@@ -1,9 +1,12 @@
 module.exports = {
-  entry: "./static/rustw.js",
+  entry: "./static/rustw.ts",
   output: {
     filename: "./static/rustw.out.js",
     libraryTarget: 'var',
     library: 'Rustw'
+  },
+  resolve: {
+    extensions: [".js", ".ts", ".tsx"]
   },
   module: {
     loaders: [
@@ -11,6 +14,11 @@ module.exports = {
       test: /\.js$/,
       exclude: /node_modules/,
       loader: 'babel-loader'
+    },
+    {
+      test: /\.tsx?$/,
+      exclude: /node_modules/,
+      loader: 'ts-loader'
     }]
   },
   devtool: 'source-map'
